DrawLetters overhaul
Changed from using pointArray to a switch case block so that we can just update the points we need as we need them.
Created a LetterLines(int,int, bool) function that is essentially DrawLines() but it will automatically update the given points before drawing the line.

DrawA-I completed.

Also went through and got everything to compile.